+++ tests/testthat/test-read-fwf.R
@@ -133,3 +133,8 @@ test_that("read_table silently reads ragged last column", {
   x <- read_table("foo bar\n1   2\n3   4\n5   6\n", progress = FALSE)
   expect_equal(x$foo, c(1, 3, 5))
 })
+
+test_that("read_table can read from a pipe (552)", {
+  x <- read_table(pipe("echo a b c && echo 1 2 3 && echo 4 5 6"))
+  expect_equal(x$a, c(1, 4))
+})